django-pdf/requirements.txt
django==4.1.3
uwsgi
pytest-django==4.5.2
psycopg2==2.9.5
pandas==1.5.1
faker==15.3.1
boto3==1.26.4
countryinfo==0.1.2
django-redis==5.2.0
django-bootstrap-v5==1.0.11
django-tables2==2.4.1
django-weasyprint==2.2.0
reportlab==3.6.12
selenium==4.6.0
seaborn